Refactoring in progress...
