+++ components/Form/index.js
@@ -1,6 +1,4 @@
-import mongoose from "mongoose";
 import { formatDateForInput } from "@/lib/utils";
-import { useState } from "react";
 import {
   ButtonContainer,
   StyledTextButton,
@@ -34,6 +32,7 @@ import {
 import { useFormData } from "@/components/Form/Form.handlers";
 import ImageUpload from "@/components/ImageUpload";
 import PresetSelect from "@/components/PresetSelect";
+import PackingList from "@/components/PackingList";
 
 const INITIAL_DATA = {
   destination: "",
@@ -66,149 +65,6 @@ export default function Form({
     handleReset,
     handleSubmit,
   } = useFormData(defaultData, onSubmit, isEditMode);
-
-  const [newPackingListItem, setNewPackingListItem] = useState({
-    itemName: "",
-    itemQuantity: null,
-  });
-  const [selectedPresetData, setSelectedPresetData] = useState([]);
-  const [lastAppliedPreset, setLastAppliedPreset] = useState(null);
-
-  function generateObjectId() {
-    const { ObjectId } = mongoose.Types;
-    const newObjectId = new ObjectId().toString();
-    return newObjectId;
-  }
-
-  function handleAddItem() {
-    if (formDisabled) {
-      return;
-    }
-
-    const lastItem =
-      handoverData.packingList[handoverData.packingList.length - 1];
-
-    if (lastItem && lastItem.itemName === "") {
-      return;
-    }
-
-    const nextPackingListItem = {
-      ...newPackingListItem,
-      _id: generateObjectId(),
-    };
-
-    const updatedPackingList = [
-      ...handoverData.packingList,
-      nextPackingListItem,
-    ];
-
-    setHandoverData((prevData) => ({
-      ...prevData,
-      packingList: updatedPackingList,
-    }));
-
-    setNewPackingListItem({ itemName: "", itemQuantity: null });
-    setHasChanges(true);
-  }
-
-  function handleRemoveItem(itemIdToRemove) {
-    if (formDisabled) {
-      return;
-    }
-
-    setHandoverData((prevData) => {
-      const updatedPackingList = handoverData.packingList.filter((item) => {
-        return item._id !== itemIdToRemove;
-      });
-
-      return {
-        ...prevData,
-        packingList: updatedPackingList,
-      };
-    });
-    setHasChanges(true);
-  }
-
-  function handleUpdateItem(itemId, itemName, itemQuantity) {
-    setHandoverData((prev) => {
-      const updatedPackingList = prev.packingList.map((item) =>
-        item._id === itemId ? { ...item, itemName, itemQuantity } : item
-      );
-
-      return {
-        ...prev,
-        packingList: updatedPackingList,
-      };
-    });
-
-    setHasChanges(true);
-  }
-
-  function handleUpdateNewItemName(
-    newName,
-    newPackingListItem,
-    setNewPackingListItem
-  ) {
-    const updatedNewPackingListItem = {
-      itemName: newName,
-      itemQuantity: newPackingListItem.itemQuantity,
-    };
-    setNewPackingListItem(updatedNewPackingListItem);
-  }
-
-  function handleUpdateNewItemQuantity(
-    newQuantity,
-    newPackingListItem,
-    setNewPackingListItem
-  ) {
-    const updatedNewPackingListItem = {
-      itemQuantity: newQuantity,
-      itemName: newPackingListItem.itemName,
-    };
-    setNewPackingListItem(updatedNewPackingListItem);
-  }
-
-  function generateListFromPreset(selectedPresetData) {
-    console.log("Apply:", selectedPresetData);
-    const selectedPreset = selectedPresetData.preset;
-
-    if (!selectedPreset) {
-      toast.error("No preset selected yet.", {
-        duration: toastDuration,
-      });
-      return;
-    }
-    if (lastAppliedPreset === selectedPreset) {
-      return;
-    }
-    setLastAppliedPreset(selectedPreset);
-
-    const updatedPackingList = [...handoverData.packingList];
-
-    const lastItem = updatedPackingList[updatedPackingList.length - 1];
-    if (lastItem && lastItem.itemName === "") {
-      updatedPackingList.pop();
-      updatedPackingList.push(
-        ...selectedPresetData.items.map((item) => ({
-          ...item,
-          _id: generateObjectId(),
-        }))
-      );
-    } else {
-      updatedPackingList.push(
-        ...selectedPresetData.items.map((item) => ({
-          ...item,
-          _id: generateObjectId(),
-        }))
-      );
-    }
-
-    setHandoverData((prevData) => ({
-      ...prevData,
-      packingList: updatedPackingList,
-    }));
-    setHasChanges(true);
-  }
 
   return (
     <TripForm
@@ -291,6 +147,7 @@ export default function Form({
           disabled={formDisabled}
         />
       )}
+
       <PackListContainer disabled={formDisabled}>
         <StyledLabel htmlFor="packingList">Packing List</StyledLabel>
         <TemplateContainer>
@@ -308,7 +165,7 @@ export default function Form({
             Apply
           </StyledTextButtonMediumSize>
         </TemplateContainer>
-        <PackList>
+        {/* <PackList>
           {handoverData.packingList.length > 0 && (
             <ItemHeaderLabel>
               <ItemNumberLabel>No.</ItemNumberLabel>
@@ -349,8 +206,8 @@ export default function Form({
               +
             </StyledMiniButton>
             <MiniButtonLabel>Add Packing List Item</MiniButtonLabel>
-          </MiniButtonContainer>{" "}
-        </PackList>
+          </MiniButtonContainer>
+        </PackList> */}
       </PackListContainer>
 
       <StyledLabel htmlFor="notes">Notes</StyledLabel>
@@ -375,74 +232,5 @@ export default function Form({
         </StyledTextButton>
       </ButtonContainer>
     </TripForm>
-  );
-}
-
-function InputItemAndQuantity({
-  item,
-  handleUpdateItem,
-  handleRemoveItem,
-  formDisabled,
-}) {
-  return (
-    <>
-      <InputItemName
-        id={`packingList_${item._id}`}
-        name={`packingList_${item._id}`}
-        type="text"
-        value={item.itemName}
-        onChange={(event) =>
-          handleUpdateItem(item._id, event.target.value, item.itemQuantity)
-        }
-        disabled={formDisabled}
-      />
-      <InputItemQuantity
-        id={`packingList_quantity_${item._id}`}
-        name={`packingList_quantity_${item._id}`}
-        type="number"
-        value={item.itemQuantity}
-        onChange={(event) =>
-          handleUpdateItem(item._id, item.itemName, event.target.value)
-        }
-        disabled={formDisabled}
-        min="0"
-        max="999"
-      />
-      <StyledMiniButton
-        type="button"
-        id="delete"
-        action="delete"
-        onClick={() => handleRemoveItem(item._id)}
-        disabled={formDisabled}
-      >
-        X
-      </StyledMiniButton>
-    </>
-  );
-}
-
-function NewPackingListItem({
-  newPackingListItem,
-  handleUpdateNewItemName,
-  handleUpdateNewItemQuantity,
-  formDisabled,
-}) {
-  return (
-    <InputContainer>
-      <InputItemName
-        type="text"
-        disabled={formDisabled}
-        value={newPackingListItem.itemName}
-        onChange={(event) => handleUpdateNewItemName(event.target.value)}
-      />
-      <InputItemQuantity
-        type="number"
-        disabled={formDisabled}
-        value={newPackingListItem.itemQuantity}
-        onChange={(event) => handleUpdateNewItemQuantity(event.target.value)}
-        min="0"
-        max="999"
-      />
-    </InputContainer>
   );
 }
